ai-rag-weather/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-rag-weather"
version = "0.1.0"
description = "RAG + WeatherBot: LangChain, LangGraph, Qdrant, Streamlit, LangSmith"
authors = [
    { name = "Surya" }
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.16",
    "langgraph>=0.0.38",
    "qdrant-client>=1.7.3",
    "sentence-transformers>=2.2.2",
    "streamlit>=1.32.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.4",
    "tenacity>=8.2.3",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "black>=24.4.0",
    "pre-commit>=3.7.0",
    "types-requests",
    "types-python-dateutil"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "C", "W", "UP", "N", "A", "S", "T", "Q"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

